Use new preact compat (#4914)
