+++ packages/app/src/sandbox/eval/index.ts
@@ -15,11 +15,12 @@ import {
 } from '@codesandbox/common/lib/templates';
 
 import { isBabel7 } from '@codesandbox/common/lib/utils/is-babel-7';
+import { isPreact10 } from '@codesandbox/common/lib/utils/is-preact-10';
 import { PackageJSON } from '@codesandbox/common/lib/types';
 import { reactPresetV1, reactPresetV3 } from './presets/create-react-app';
 import reactTsPreset from './presets/create-react-app-typescript';
 import vuePreset from './presets/vue-cli';
-import preactPreset from './presets/preact-cli';
+import { preactPreset, preactPresetV8 } from './presets/preact-cli';
 import sveltePreset from './presets/svelte';
 import angularPreset from './presets/angular-cli';
 import parcelPreset from './presets/parcel';
@@ -38,14 +39,20 @@ export default function getPreset(template: string, pkg: PackageJSON) {
 
       return reactPresetV1();
 
+    case preact.name:
+      if (isPreact10(pkg.dependencies, pkg.devDependencies)) {
+        return preactPreset();
+      }
+
+      return preactPresetV8();
+
     case reactTs.name:
       return reactTsPreset();
     case reason.name:
       return reasonPreset();
     case vue.name:
       return vuePreset();
-    case preact.name:
-      return preactPreset();
+
     case svelte.name:
       return sveltePreset();
     case angular.name:
